Create document class
