+++ include/CMakeLists.txt
@@ -1,6 +1,8 @@
 set(SIMDJSON_INCLUDE_DIR ${PROJECT_SOURCE_DIR}/include)
 set(SIMDJSON_INCLUDE
     ${SIMDJSON_INCLUDE_DIR}/simdjson/common_defs.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/document.h
+    ${SIMDJSON_INCLUDE_DIR}/simdjson/document/iterator.h
     ${SIMDJSON_INCLUDE_DIR}/simdjson/isadetection.h
     ${SIMDJSON_INCLUDE_DIR}/simdjson/jsonformatutils.h
     ${SIMDJSON_INCLUDE_DIR}/simdjson/jsonioutil.h
@@ -8,7 +10,6 @@ set(SIMDJSON_INCLUDE
     ${SIMDJSON_INCLUDE_DIR}/simdjson/jsonparser.h
     ${SIMDJSON_INCLUDE_DIR}/simdjson/padded_string.h
     ${SIMDJSON_INCLUDE_DIR}/simdjson/parsedjson.h
-    ${SIMDJSON_INCLUDE_DIR}/simdjson/parsedjsoniterator.h
     ${SIMDJSON_INCLUDE_DIR}/simdjson/portability.h
     ${SIMDJSON_INCLUDE_DIR}/simdjson/simdjson.h
     ${SIMDJSON_INCLUDE_DIR}/simdjson/simdjson_version.h
