+++ src/generic/stage1_find_marks.h
@@ -392,33 +392,33 @@ really_inline void json_structural_scanner::scan(const uint8_t *buf, const size_
 // you may want to call on a function like trimmed_length_safe_utf8.
 template<size_t STEP_SIZE>
 int find_structural_bits(const uint8_t *buf, size_t len, simdjson::ParsedJson &pj, bool streaming) {
-  if (unlikely(len > pj.byte_capacity)) {
+  if (unlikely(len > pj.doc.byte_capacity)) {
     return simdjson::CAPACITY;
   }
   utf8_checker utf8_checker{};
-  json_structural_scanner scanner{pj.structural_indexes.get()};
+  json_structural_scanner scanner{pj.doc.structural_indexes.get()};
   scanner.scan<STEP_SIZE>(buf, len, utf8_checker);
   // we might tolerate an unclosed string if streaming is true
   simdjson::ErrorValues error = scanner.detect_errors_on_eof(streaming);
   if (unlikely(error != simdjson::SUCCESS)) {
     return error;
   }
-  pj.n_structural_indexes = scanner.structural_indexes.tail - pj.structural_indexes.get();
+  pj.doc.n_structural_indexes = scanner.structural_indexes.tail - pj.doc.structural_indexes.get();
   /* a valid JSON file cannot have zero structural indexes - we should have
    * found something */
-  if (unlikely(pj.n_structural_indexes == 0u)) {
+  if (unlikely(pj.doc.n_structural_indexes == 0u)) {
     return simdjson::EMPTY;
   }
-  if (unlikely(pj.structural_indexes[pj.n_structural_indexes - 1] > len)) {
+  if (unlikely(pj.doc.structural_indexes[pj.doc.n_structural_indexes - 1] > len)) {
     return simdjson::UNEXPECTED_ERROR;
   }
-  if (len != pj.structural_indexes[pj.n_structural_indexes - 1]) {
+  if (len != pj.doc.structural_indexes[pj.doc.n_structural_indexes - 1]) {
     /* the string might not be NULL terminated, but we add a virtual NULL
      * ending character. */
-    pj.structural_indexes[pj.n_structural_indexes++] = len;
+    pj.doc.structural_indexes[pj.doc.n_structural_indexes++] = len;
   }
   /* make it safe to dereference one beyond this array */
-  pj.structural_indexes[pj.n_structural_indexes] = 0;
+  pj.doc.structural_indexes[pj.doc.n_structural_indexes] = 0;
   return utf8_checker.errors();
 }
