--- src/parsedjsoniterator.cpp
@@ -1,5 +1,0 @@
-#include "simdjson/parsedjsoniterator.h"
-
-namespace simdjson {
-template class ParsedJson::BasicIterator<DEFAULT_MAX_DEPTH>;
-} // namespace simdjson
